api/requirements.txt
Flask==2.2.5
python-telegram-bot==20.3
requests==2.31.0
